Replace icons for France Bleu stations
Many of the station icons for France Bleu stations were not being shown in LMS.
This is because around half of the France Bleu .png/.jpg station icons have disappeared from the FranceBleu.fr site and they have switched to using .svg instead.

LMS and many players do not like .svg images (Material skin is OK with them) so rather than link directly to the .svg images on the remote web site, I now include .svg and .png (made by exporting from the .svg) images in the plugin for these FB stations (apart from one that is not available as .svg).